Приложение обновлено для работы с БД согласно заданию.

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmStorage.java
@@ -11,5 +11,5 @@ public interface FilmStorage {
 
     Film update(Film newFilm);
 
-    Film getFilm(Long filmId);
+    Film getFilm(Integer filmId);
 }
